Core/Core-GTK: fs_unlink() shouldn't raise an exception if the file doesn't exist

+++ openpaperwork-core/src/openpaperwork_core/fs/python.py
@@ -70,6 +70,8 @@ class Plugin(CommonFsPluginBase):
         if path is None:
             return
 
+        if not os.path.exists(path):
+            return None
         os.unlink(path)
         return True
